signed/unsigned comparison error.

+++ wire/Message.h
@@ -69,13 +69,16 @@ static inline void Message_copyOver(struct Message* output,
 static inline bool Message_shift(struct Message* toShift, int32_t amount)
 {
     Assert_true(toShift->padding >= amount);
-    Assert_true(
-        (amount >= 0) ? (UINT16_MAX - toShift->length >= amount) : (toShift->length >= -amount));
-Assert_true(toShift->length < 60000);
+    Assert_true((amount >= 0) ? (((int32_t)UINT16_MAX) - toShift->length >= amount)
+                              : (((int32_t)toShift->length) >= -amount));
+    Assert_true(toShift->length < 60000);
+
     toShift->length += amount;
     toShift->bytes -= amount;
     toShift->padding -= amount;
-Assert_true(toShift->length < 60000);
+
+    Assert_true(toShift->length < 60000);
+
     return true;
 }
